Improvements to stage speed. Enemy spawn object changes.
- Enemy spawn objects will now instantiate all objects instantly instead of on a delay.
- Enemy spawn objects will now spawn extra copies in one of four directions (left, right, up or down). This direction can be set in the inspector.

- Added a new script called the enemy encounter queue. 
This script is attached to the enemy spawner group object and will determine if the game needs to fast forward to the next enemy spawner object. 
This fixes a game play issue where the player is stuck waiting for game to spawn the next set of enemies if they were able to destroy all enemies too quickly.
Now the next set of enemies will spawn if the amount of active enemies is 0 and one second passes.
For debugging purposes, you can press the 'H' key to immediately go to the next encounter.